backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlmodel==0.0.16
alembic==1.13.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
markdown-it-py==3.0.0
bleach==6.1.0
slowapi==0.1.9
python-dateutil==2.9.0.post0
pydantic[email]==1.10.15
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.5
